Removed creating hands with cards, removed draw redundancy, fixed card_count
Whenever a hand was initialised, you could pass a list of cards to it to create that hand with cards already in it, this functionality was removed as it conflicted with the drawing and dealing of cards, aswell as it was redundant to the fact that decks should be controlling hands creation. Also, the required argument of deck was removed from the draw function, this was redundant as hands are already created with a local deck property as the instance of the deck that created them. Also, the card_count property had inconsistancies with the naming conventions used for it. The card_count function should now accurately reflect the size of the hand whenever it is called.